requirements.txt
# Core dependencies
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
scipy>=1.7.0,<2.0.0

# Deep Learning
tensorflow>=2.10.0,<2.15.0
keras>=2.10.0,<3.0.0

# Visualization
matplotlib>=3.4.0,<4.0.0
seaborn>=0.11.0,<1.0.0

# Configuration and utilities
pyyaml>=5.4.0,<7.0.0
tqdm>=4.62.0,<5.0.0

# Jupyter support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Development tools (optional)
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.6b0
flake8>=3.9.0
mypy>=0.910

# Additional utilities
h5py>=3.0.0  # For saving Keras models
tensorboard>=2.10.0  # For TensorBoard visualization
protobuf>=3.19.0,<4.0.0  # Required by TensorFlow